Removing config-class, using configs directly in the stack to minimize repetition and simplify code.
